requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0

tensorflow-cpu==2.20.0
numpy
python-dotenv
